Reapply changes to fix bad rebase.
