+++ packages/core-data/README.md
@@ -976,6 +976,47 @@ The following set of react hooks available to import from the `@wordpress/core-d
 
 <!-- START TOKEN(Autogenerated hooks|src/hooks/index.ts) -->
 
+### useEntityBlockEditor
+
+Hook that returns block content getters and setters for the nearest provided entity of the specified type.
+
+The return value has the shape `[ blocks, onInput, onChange ]`. `onInput` is for block changes that don't create undo levels or dirty the post, non-persistent changes, and `onChange` is for persistent changes. They map directly to the props of a `BlockEditorProvider` and are intended to be used with it, or similar components or hooks.
+
+_Parameters_
+
+-   _kind_ `string`: The entity kind.
+-   _name_ `string`: The entity name.
+-   _options_ `Object`:
+-   _options.id_ `[string]`: An entity ID to use instead of the context-provided one.
+
+_Returns_
+
+-   `[unknown[], Function, Function]`: The block array and setters.
+
+### useEntityId
+
+Hook that returns the ID for the nearest provided entity of the specified type.
+
+_Parameters_
+
+-   _kind_ `string`: The entity kind.
+-   _name_ `string`: The entity name.
+
+### useEntityProp
+
+Hook that returns the value and a setter for the specified property of the nearest provided entity of the specified type.
+
+_Parameters_
+
+-   _kind_ `string`: The entity kind.
+-   _name_ `string`: The entity name.
+-   _prop_ `string`: The property name.
+-   _\_id_ `[string]`: An entity ID to use instead of the context-provided one.
+
+_Returns_
+
+-   `[*, Function, *]`: An array where the first item is the property value, the second is the setter and the third is the full value object from REST API containing more information like `raw`, `rendered` and `protected` props.
+
 ### useEntityRecord
 
 Resolves the specified entity record.
