+++ packages/core-data/src/hooks/index.ts
@@ -10,3 +10,6 @@ export {
 	default as useResourcePermissions,
 	__experimentalUseResourcePermissions,
 } from './use-resource-permissions';
+export { default as useEntityBlockEditor } from './use-entity-block-editor';
+export { default as useEntityId } from './use-entity-id';
+export { default as useEntityProp } from './use-entity-prop';
